fix(plugin): follow the documented test filename and refresh stale SKILL.md docs
Review follow-ups on the snapshot backfill, both against rules in
hvantk/skills/_conventions/SKILL.md.

Section 8 states the round-trip test file is `test_builder.py` for Hail-backed
providers. All seven new files were named `test_builder_snapshot.py`; none of the
seven plugins had a pre-existing `test_builder.py`, so nothing forced the
deviation. Renamed, and the self-referential --regenerate-snapshots hint in each
docstring updated to match. The plugin-contract CI job selects by directory, so
it picks the new names up without a workflow change.

Section 9 requires each SKILL.md to declare paths that match the `tests:` block in
plugin.yaml. Five were left asserting things this branch had made false:

  * clingen, gencc, hgnc still said snapshots "have NOT yet been seeded" -- this
    branch seeds them. gencc additionally pointed at test_gencc.py to bootstrap
    them, which never held the regenerate call.
  * dbnsfp and gnomad-metrics still described a plugin-relative fixture after
    plugin.yaml was repointed at the shared hvantk/tests/testdata/raw/ location.

The dbnsfp and gnomad-metrics entries now state the split explicitly -- snapshots
and drift fingerprint are plugin-relative, the raw fixture is not -- and say why
it is referenced in place rather than duplicated, so the deviation from the
plugin-relative convention is documented rather than silent.

Verified on Slurm (Hail 0.2.137, Java 11): the CI selection is 56 passed / 0
failed after the rename, with all 12 test_builder.py round-trips green; the
ratchet is unchanged at 14 passed.
