+++ hvantk/skills/ensembl_gene/tests/test_builder.py
@@ -1,11 +1,11 @@
-"""Snapshot round-trip test for the dbNSFP builder.
+"""Snapshot round-trip test for the Ensembl gene builder.
 
 Builds from the committed fixture and asserts the schema and a sample of rows against
 committed snapshots, so a change in builder behaviour or upstream field layout shows up
 as an explicit diff rather than silently.
 
 Regenerate after an intentional change:
-    pytest hvantk/skills/dbnsfp/tests/test_builder_snapshot.py --regenerate-snapshots
+    pytest hvantk/skills/ensembl_gene/tests/test_builder.py --regenerate-snapshots
 """
 
 from __future__ import annotations
@@ -23,28 +23,28 @@ from hvantk.tests._snapshot_utils import (
 # Aliased to avoid shadowing the fixture name `regenerate_snapshots` in the test signature
 from hvantk.tests._snapshot_utils import regenerate_snapshots as regenerate_snapshots_fn
 
-FIXTURE = "hvantk/tests/testdata/raw/dbnsfp/dbNSFP4_v49a_example_variants.bgz"
-SNAPSHOT_DIR = Path("hvantk/skills/dbnsfp/tests/snapshots")
+FIXTURE = "hvantk/tests/testdata/raw/ensembl/ensembl_gene_biomart.tsv.bgz"
+SNAPSHOT_DIR = Path("hvantk/skills/ensembl_gene/tests/snapshots")
 
 # Real keys, taken from an actual build of the fixture -- collect_sample_rows raises
 # KeyError for any key absent from the table, so these cannot be invented.
 SAMPLE_KEYS = [
-    {"alleles": ["C", "A"], "locus": "chr10:47057"},
-    {"alleles": ["C", "G"], "locus": "chr10:47057"},
-    {"alleles": ["T", "A"], "locus": "chr10:47058"},
-    {"alleles": ["T", "C"], "locus": "chr10:47058"},
-    {"alleles": ["T", "G"], "locus": "chr10:47058"},
-    {"alleles": ["A", "C"], "locus": "chr10:47059"},
+    {"gene_id": "ENSG00000000003"},
+    {"gene_id": "ENSG00000000005"},
+    {"gene_id": "ENSG00000000419"},
+    {"gene_id": "ENSG00000000457"},
+    {"gene_id": "ENSG00000000460"},
+    {"gene_id": "ENSG00000000938"},
 ]
 
 
 @pytest.mark.hail
-def test_dbnsfp_snapshot_round_trip(hail_session, tmp_path, regenerate_snapshots):
-    """Build dbNSFP from fixture; assert schema and sample-row stability."""
+def test_ensembl_gene_snapshot_round_trip(hail_session, tmp_path, regenerate_snapshots):
+    """Build Ensembl gene from fixture; assert schema and sample-row stability."""
     import hail as hl
-    from hvantk.skills.dbnsfp.builder import build_dbnsfp_variants
+    from hvantk.skills.ensembl_gene.builder import build_ensembl_gene_genes
 
-    builder = phase_b_snapshot_adapter(build_dbnsfp_variants, "dbnsfp:variants")
+    builder = phase_b_snapshot_adapter(build_ensembl_gene_genes, "ensembl-gene:genes")
 
     if regenerate_snapshots:
         regenerate_snapshots_fn(
@@ -55,13 +55,13 @@ def test_dbnsfp_snapshot_round_trip(hail_session, tmp_path, regenerate_snapshots
         )
         pytest.skip("Snapshots regenerated; rerun without --regenerate-snapshots to assert.")
 
-    output_path = str(tmp_path / "dbnsfp.ht")
+    output_path = str(tmp_path / "ensembl_gene.ht")
     builder(input_path=FIXTURE, output_path=output_path)
     ht = hl.read_table(output_path)
 
     expected_schema = load_snapshot(SNAPSHOT_DIR / "schema.json")
-    assert hail_schema_to_dict(ht) == expected_schema, "dbNSFP schema drifted from snapshot"
+    assert hail_schema_to_dict(ht) == expected_schema, "Ensembl gene schema drifted from snapshot"
 
     expected_rows = load_snapshot(SNAPSHOT_DIR / "sample_rows.json")
     actual_rows = collect_sample_rows(ht, keys=SAMPLE_KEYS)
-    assert actual_rows == expected_rows, "dbNSFP sample rows drifted from snapshot"
+    assert actual_rows == expected_rows, "Ensembl gene sample rows drifted from snapshot"
