+++ hvantk/skills/gencc/SKILL.md
@@ -82,8 +82,8 @@ When GenCC publishes an updated snapshot (any download is effectively a new snap
 - `drift_fingerprint`: `hvantk/skills/gencc/tests/drift_fingerprint.json`
 - `test_command`: `pytest hvantk/skills/gencc/tests -m hail`
 
-> **Snapshot status:** schema.json and sample_rows.json have NOT yet been seeded
-> for this plugin. On first round-trip run in a hail-enabled environment, use
-> `pytest hvantk/skills/gencc/tests/test_gencc.py --regenerate-snapshots`
-> to bootstrap them, then commit. Until seeded, the round-trip test cannot verify
-> output against a fixed schema.
+> **Snapshot status:** seeded. `tests/snapshots/schema.json` and
+> `tests/snapshots/sample_rows.json` are committed, and `test_builder.py` asserts the
+> build against them. Regenerate after an intentional schema change with
+> `pytest hvantk/skills/gencc/tests/test_builder.py --regenerate-snapshots`, then
+> commit the result.
